model pipeline changes

+++ src/api/schema.py
@@ -8,18 +8,18 @@ class HeartDiseaseFeatures(BaseModel):
     Based on the columns expected by the model.
     """
     age: float = Field(..., description="Age in years.")
-    sex: float = Field(..., description="Sex (1 = male; 0 = female).")
-    cp: float = Field(..., description="Chest pain type (1-4, mapped to categories).")
+    sex: int = Field(..., description="Sex (1 = male; 0 = female).")
+    cp: int = Field(..., description="Chest pain type (1-4, mapped to categories).")
     trestbps: float = Field(..., description="Resting blood pressure (in mm Hg on admission to the hospital).")
     chol: float = Field(..., description="Serum cholestoral in mg/dl.")
-    fbs: float = Field(..., description="Fasting blood sugar > 120 mg/dl (1 = true; 0 = false).")
-    restecg: float = Field(..., description="Resting electrocardiographic results (0, 1, or 2).")
+    fbs: int = Field(..., description="Fasting blood sugar > 120 mg/dl (1 = true; 0 = false).")
+    restecg: int = Field(..., description="Resting electrocardiographic results (0, 1, or 2).")
     thalach: float = Field(..., description="Maximum heart rate achieved.")
-    exang: float = Field(..., description="Exercise induced angina (1 = yes; 0 = no).")
+    exang: int = Field(..., description="Exercise induced angina (1 = yes; 0 = no).")
     oldpeak: float = Field(..., description="ST depression induced by exercise relative to rest.")
-    slope: float = Field(..., description="The slope of the peak exercise ST segment (1, 2, or 3).")
-    ca: float = Field(..., description="Number of major vessels (0-3) colored by flourosopy.")
-    thal: float = Field(..., description="Thalium stress test result (3 = normal; 6 = fixed defect; 7 = reversable defect).")
+    slope: int = Field(..., description="The slope of the peak exercise ST segment (1, 2, or 3).")
+    ca: int = Field(..., description="Number of major vessels (0-3) colored by flourosopy.")
+    thal: int = Field(..., description="Thalium stress test result (3 = normal; 6 = fixed defect; 7 = reversable defect).")
     
     # Example for documentation and validation
     model_config = {
@@ -27,18 +27,18 @@ class HeartDiseaseFeatures(BaseModel):
             "examples": [
                 {
                     "age": 63.0,
-                    "sex": 1.0,
-                    "cp": 1.0,
+                    "sex": 1,
+                    "cp": 1,
                     "trestbps": 145.0,
                     "chol": 233.0,
-                    "fbs": 1.0,
-                    "restecg": 2.0,
+                    "fbs": 1,
+                    "restecg": 2,
                     "thalach": 150.0,
-                    "exang": 0.0,
+                    "exang": 0,
                     "oldpeak": 2.3,
-                    "slope": 3.0,
-                    "ca": 0.0,
-                    "thal": 6.0
+                    "slope": 3,
+                    "ca": 0,
+                    "thal": 6
                 }
             ]
         }
